resized windows

+++ a4/Box.pde
@@ -5,7 +5,7 @@ public class Box {
     private ArrayList<Edge> edges;
     private float weight;
     public int index;
-    public final int margin_l = 50, margin_b = 20;
+    public final float margin_l = 50, margin_b = 20;
     public Boolean selected;
     public float x, y, w, h;
